requirements.txt
numpy
labjack-ljm
StrEnum